Added pthread_join, ft_sleep, organised functions, philo[i] 1 -> 0

+++ includes/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: mialbert <mialbert@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/10 00:42:58 by mialbert          #+#    #+#             */
-/*   Updated: 2022/09/16 00:08:10 by mialbert         ###   ########.fr       */
+/*   Updated: 2022/09/17 23:06:09 by mialbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ typedef struct data
 {
 	t_philo			*philo;
 	pthread_mutex_t	*forks;
-	size_t			philo_nbr;
+	int32_t			philo_nbr;
 	int32_t			ripoclock;
 	int32_t			nomoclock;
 	int32_t			zzzoclock;
@@ -37,6 +37,7 @@ typedef struct data
 typedef struct philo
 {
 	pthread_t		sopher;
+	
 	int32_t			index;
 	int64_t			meal_time;
 	int32_t			meal_count;
